made metadata parser more robust and moved to own file
