+++ djweb/dj_comp_hist/models.py
@@ -190,136 +190,3 @@ def interpret_person_organization(field, item_organization, item_person, new_doc
             bound_attr = getattr(new_doc, item_person)
             bound_attr.add(new_item)
 
-
-def populate_from_metadata(file_name=None):
-    '''
-    :param file_name: is a path to a csv file (relative or actual)
-    :return: populates the django database, but returns nothing
-
-    how to run:
-    open up terminal (with virtual environment):
-    > cd djweb
-    > python manage.py shell
-
-    in shell:
-    > from dj_comp_hist.models import populate_from_metadata
-    > populate_from_metadata()
-
-    The 'r' in front of the file location isn't necessary but can help to prevent strange errors.
-    Utilizes interpret_organization_person to add authors, recipients, and cced.
-    '''
-
-    if file_name is None:
-        file_name = Path(os.path.abspath(os.path.dirname(__file__)), 'metadata_jan24.csv')
-
-    with open(file_name) as file:
-        csv_file = csv.DictReader(file)
-        for line in csv_file:
-            number_of_pages = int(line['last_page']) - int(line['first_page']) + 1
-            new_doc = Document(number_of_pages=number_of_pages,
-                               title=line['title'],
-                               type=line['doc_type'],
-                               notes=line['notes'],
-                               first_page=line['first_page'],  # should this be an int?
-                               last_page=line['last_page'],
-                               file_name=line['filename']
-                               )
-
-            # ---------------------DATE-----------------------------------------------
-            if line['date'] != '' and line['date'][0] == '1':
-                new_doc.date = line['date']
-
-            # ------------------------------------------------------------------------
-
-            # ---------------------Folder---------------------------------------------
-            folder_exist, new_folder = check_generate(Folder, "name", line['foldername_short'])
-            if not folder_exist:
-                box_exist, new_box = check_generate(Box, "number", line['box'])
-                new_box.save()
-                new_folder.box = new_box
-                new_folder.number = line['folder_number']
-                new_folder.full = line['foldername_full']
-                new_folder.file_name = line['filename']
-            new_folder.save()
-            new_doc.folder = new_folder
-            new_doc.save()
-
-            # -----------------------Author, Recipient,cced--------------------------
-            interpret_person_organization(line['author'],
-                                          "author_organization",
-                                          "author_person",
-                                          new_doc
-                                          )
-            interpret_person_organization(line['recipients'],
-                                          "recipient_organization",
-                                          "recipient_person",
-                                          new_doc
-                                          )
-            interpret_person_organization(line['cced'],
-                                          "cced_organization",
-                                          "cced_person",
-                                          new_doc
-                                          )
-
-            new_doc.save()
-        return
-
-
-def pdf_to_image_split(pdf_path, image_directory, folder_name):
-    """
-    Splits a each page of a pdf into an image.
-    pdf_path is the location of pdf (C:\Documents\rockefeller.pdf)
-    image_directory is the location of image folder (C:\Documents\png_pages\)
-    folder_name is the names of the object of the folder(rockefeller)
-    """
-    pages = convert_from_path(pdf_path)
-    images_in_pdf = []
-
-    for page in range(len(pages)):
-        page_path = image_directory + folder_name + '_' + str(page + 1) + '.png'
-        pages[page].save(page_path, 'PNG')
-        images_in_pdf.append((page_path, page+1))
-
-    return images_in_pdf
-
-
-def page_image_to_doc(folder_name, pdf_path, image_directory):
-    """
-    Utilizes pdf_to_image_split in order to create images of Pages from a pdf, create Page
-    objects, and assign those page objects to Document.
-    Splits a each page of a pdf into an image.
-    pdf_path is the location of pdf (C:\Documents\rockefeller.pdf)
-    image_directory is the location of image folder (C:\Documents\png_pages\)
-    folder_name is the names of the object of the folder(rockefeller)
-    Example: page_image_to_doc('rockefeller', 'C:\Documents\1_08_raw_rockefeller.pdf',
-    'C:\Documents\png_pages\') returns images of Pages in png_pages directory and Page objects
-    of each page.
-    """
-    images_in_pdf = pdf_to_image_split(pdf_path, image_directory, folder_name)
-    folder = Folder.objects.get(name=folder_name)
-    documents_unsort = folder.document_set.all()
-    documents_sort = sorted(documents_unsort, key=lambda x: x.first_page)
-    document_place = 0
-    page_num = 1
-
-    for page in images_in_pdf:
-        if documents_sort[document_place].first_page <= page[1] <= documents_sort[\
-                document_place].last_page:
-            # this means that this is the same document as last page
-            page_obj = Page(document=documents_sort[document_place], page_number=page_num)
-            page_obj.image_path.name = folder_name + '_' + str(page[1]) + '.png'
-            page_obj.save()
-            page_num += 1
-        elif documents_sort[document_place+1].first_page <= page[1] <= documents_sort[\
-                document_place+1].last_page:
-            # this means this is a new document
-            document_place += 1
-            page_num = 1
-            page_obj = Page(document=documents_sort[document_place], page_number=page_num)
-            page_obj.image_path.name = folder_name + '_' + str(page[1]) + '.png'
-            page_obj.save()
-            page_num += 1
-        else:
-            # This means that the page isn't associated with a document
-            pass
-    return
